+++ src/components/SupportOBSSN/SupportOBSSN.js
@@ -1,70 +1,58 @@
-import React, { useState, useEffect } from 'react';
-
+import React from 'react';
 import { Card, CardContent, useContent, useCardState } from 'translation-helps-rcl';
 
-export default function SupportTN(props) {
+export default function SupportOBSSN(props) {
   const { title, classes, onClose, type, server, fontSize, reference, resource } = props;
-
-  const [selectedQuote, setQuote] = useState({});
   const { bookId, chapter, verse } = reference;
 
   const {
-    markdown,
     items,
     resourceStatus: { loading },
     props: { languageId },
   } = useContent({
-    verse: String(verse),
-    chapter: String(chapter),
     projectId: bookId,
     ref: resource.branch ?? 'master',
     languageId: resource.languageId ?? 'ru',
-    resourceId: 'tn',
+    resourceId: 'obs-sn',
+    verse,
+    chapter,
     owner: resource.owner ?? 'door43-catalog',
     server,
   });
 
   const {
-    state: { item, headers, itemIndex, markdownView },
-    actions: { setItemIndex, setMarkdownView },
+    state: { item, headers, itemIndex },
+    actions: { setFilters, setItemIndex },
   } = useCardState({
     items,
+    verse,
+    chapter,
+    projectId: bookId,
+    resourceId: 'sn',
   });
-
-  useEffect(() => {
-    setItemIndex(0);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [reference]);
-
-  const filterArray = ['OrigQuote', 'GLQuote', 'OccurrenceNote'];
-
+  const filterArray = ['Quote', 'Note'];
   return (
     <Card
       closeable
       title={title}
       onClose={() => onClose(type)}
-      classes={classes}
+      classes={{ ...classes, children: 'tqcard' }}
       id={type}
       items={items}
-      fontSize={fontSize}
       headers={headers}
       filters={filterArray}
+      fontSize={fontSize}
       itemIndex={itemIndex}
+      setFilters={setFilters}
       setItemIndex={setItemIndex}
-      markdownView={markdownView}
-      setMarkdownView={setMarkdownView}
     >
       <CardContent
         item={item}
-        viewMode="table"
         filters={filterArray}
-        markdown={markdown}
         fontSize={fontSize}
+        viewMode="table"
         isLoading={Boolean(loading)}
         languageId={languageId}
-        markdownView={markdownView}
-        selectedQuote={selectedQuote}
-        setQuote={setQuote}
       />
     </Card>
   );