+++ src/components/SupportOBSSQ/index.js
@@ -1,0 +1,1 @@
+export { default as SupportOBSSQ } from './SupportOBSSQ';