+++ src/components/SupportTWL/SupportTWL.js
@@ -1,29 +1,10 @@
-import React, { useState, useContext } from 'react';
+import React, { useState, useEffect } from 'react';
 
 import { Card, CardContent, useContent, useCardState } from 'translation-helps-rcl';
 
-import { AppContext, ReferenceContext } from '../../context';
-import { server } from '../../config/base';
-
-export default function SupportTW(props) {
-  const { title, classes, onClose, type } = props;
-  const {
-    state: { resourcesApp, fontSize },
-  } = useContext(AppContext);
-
-  const {
-    state: { referenceSelected },
-  } = useContext(ReferenceContext);
-
-  const { bookId, chapter, verse } = referenceSelected;
-
-  let resource = false;
-  resourcesApp.forEach((el) => {
-    if (el.name === type) {
-      resource = el;
-    }
-  });
-
+export default function SupportTWL(props) {
+  const { title, classes, onClose, type, server, fontSize, reference, resource } = props;
+  const { bookId, chapter, verse } = reference;
   const [selectedQuote, setQuote] = useState({});
   const {
     markdown,
@@ -36,7 +17,7 @@ export default function SupportTW(props) {
     projectId: bookId,
     ref: resource.branch ?? 'master',
     languageId: resource.languageId ?? 'ru',
-    resourceId: 'tw',
+    resourceId: 'twl',
     owner: resource.owner ?? 'door43-catalog',
     server,
   });
@@ -46,7 +27,16 @@ export default function SupportTW(props) {
     actions: { setFilters, setItemIndex, setMarkdownView },
   } = useCardState({
     items,
+    setQuote,
+    selectedQuote,
+    verse,
+    chapter,
+    projectId: bookId,
   });
+  useEffect(() => {
+    setItemIndex(0);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [reference]);
 
   return (
     <Card