+++ src/docs/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - migration for language selection
  - component for dialogs
  - selection of resource languages to the first launch window
+ - OBS Study Questions
+ - OBS Study Notes
+ - Translation Words Links
 
 ### Changed (for changes in existing functionality)
  - message about the inability to close the last resource
@@ -23,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - scrolling to verses in obs
  - line break if the resource name is too long
  - duplicate resources are hidden
+
 
 ## [1.3.0] - 2021-10-20